requirement.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-import-export==2.9.0
django-jasmine==0.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-richtextfield==1.6.1
django-simple-history==3.2.0
django-solo==2.0.0
djangorestframework==3.13.1
drf-yasg==1.21.4
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
json2xml==3.19.5
MarkupPy==1.14
MarkupSafe==2.1.1
more-itertools==8.12.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==22.0
Pillow==9.0.1
psycopg2==2.9.2
pycparser==2.21
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
sqlparse==0.4.2
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0